The function check_disk_full now returns True! And refactored some code
The function check_disk_full were erronously previously named as check_disk_usage
so we renamed it to the check_disk_full. The function check_disk_full was returning
the value true if the value True if the disk is proceeding the specified limit
through arguments, and False otherwise, and now we fixed this bug. At last, I have
created a wrapper function for check_disk_full to check usage of only the root.